Give the wire-facing index types a fixed width
serde maps `usize` to `serialize_u64` unconditionally
(serde_core/src/ser/impls.rs:28) -- target width does not enter into it. A
generator that reproduces the serde data model faithfully therefore has to emit
u64 for these fields, which lands in TypeScript as `bigint`. `RoutineIndex` is
threaded through most of the frontend, so that would be wide churn bought with
nothing.

Nothing is bought because the producer is riscv32: a `usize` there cannot exceed
`u32::MAX`, and postcard varints for a value that fits in 32 bits are
byte-identical whether the field is declared u32 or u64. So this is
wire-compatible with already-flashed firmware in both directions.

Fixed width is also just more honest for these particular fields.
`RoutineIndex` is bit-packed into a u16 by `to_storage_index()` with a 14-bit
index, so `usize` was never the real domain -- it was three times wider than the
values it could hold.

Internal representations stay `usize`, because they are `Vec` offsets and should
be: `RoutineExecutionContext::current_step`, `transition_to(step)`, and the
schedule store's `remove_schedule`/`update_schedule` are untouched. The casts sit
exactly at the boundary where a wire value becomes an index, or vice versa.

The worst-case-size fixture in variegated-debug-codec moves from `usize::MAX` to
`u32::MAX`, which is the same intent (maximal value for the type) and still
exercises the longest varint the field can now produce. That test asserts "must
fit and round-trip", not an exact byte count, so it holds.

Verified: both examples build for thumbv8m, `cargo check --workspace` is clean,
and the 41 variegated-debug-codec host tests pass.

+++ examples/dual-boiler/src/buttons.rs
@@ -431,7 +431,7 @@ impl ButtonEventHandler {
                         MachineCommand::CancelRoutine
                     } else {
                         defmt::info!("Button {} pressed - starting routine {}", button_idx + 1, button_idx);
-                        MachineCommand::RunRoutine(RoutineIndex::Function(button_idx), None)
+                        MachineCommand::RunRoutine(RoutineIndex::Function(button_idx as u32), None)
                     };
                     return vec![command];
                 }
